+++ app/blog/page.tsx
@@ -7,62 +7,11 @@ import {format} from "date-fns";
 import {BlogCardProps, BlogPostsProps} from "@/app/components/BlogCard/BlogCard.props";
 import {getPosts} from "@/api/posts";
 
-const BlogPage = async () => {
-    const cards: BlogCardProps[] = [
-        {
-            docsHref: 'https://github.com/xLI4NOST/easyPeak/pulls',
-            docsHrefText: 'Два года назад команда Next.js',
-            tag: 'frontend',
-            date: format(new Date(), "dd MMM yyyy"),
-            image: 'https://habrastorage.org/r/w1560/getpro/habr/upload_files/da0/b0d/2cc/da0b0d2ccd24fe7ab50820c2803d3f9d.jpg',
-            href: 'https://habr.com/ru/articles/788898/',
-            title: 'Next.js App Router. Опыт использования. Путь в будущее или поворот не туда',
-            description: 'представила новый подход к роутингу, ' +
-                'который должен был стать заменой так называемому Pages Router, вместе с тем добавив ряд принципиально нового функционала.',
-            readTime: '3 минуты'
-        },
-        {
-            docsHref: 'https://github.com/xLI4NOST/easyPeak/pulls',
-            docsHrefText: 'Два года назад команда Next.js',
-            tag: 'frontend',
-            date: format(new Date(), "dd MMM yyyy"),
-            image: 'https://habrastorage.org/r/w1560/getpro/habr/upload_files/da0/b0d/2cc/da0b0d2ccd24fe7ab50820c2803d3f9d.jpg',
-            href: 'https://habr.com/ru/articles/788898/',
-            title: 'Next.js App Router. Опыт использования. Путь в будущее или поворот не туда',
-            description: 'представила новый подход к роутингу, ' +
-                'который должен был стать заменой так называемому Pages Router, вместе с тем добавив ряд принципиально нового функционала.',
-            readTime: '3 минуты'
-        },
-        {
-            docsHref: 'https://github.com/xLI4NOST/easyPeak/pulls',
-            docsHrefText: 'Два года назад команда Next.js',
-            tag: 'frontend',
-            date: format(new Date(), "dd MMM yyyy"),
-            image: 'https://habrastorage.org/r/w1560/getpro/habr/upload_files/da0/b0d/2cc/da0b0d2ccd24fe7ab50820c2803d3f9d.jpg',
-            href: 'https://habr.com/ru/articles/788898/',
-            title: 'Next.js App Router. Опыт использования. Путь в будущее или поворот не туда',
-            description: 'представила новый подход к роутингу, ' +
-                'который должен был стать заменой так называемому Pages Router, вместе с тем добавив ряд принципиально нового функционала.',
-            readTime: '3 минуты'
-        },
-        {
-            docsHref: 'https://github.com/xLI4NOST/easyPeak/pulls',
-            docsHrefText: 'Два года назад команда Next.js',
-            tag: 'frontend',
-            date: format(new Date(), "dd MMM yyyy"),
-            image: 'https://habrastorage.org/r/w1560/getpro/habr/upload_files/da0/b0d/2cc/da0b0d2ccd24fe7ab50820c2803d3f9d.jpg',
-            href: 'https://habr.com/ru/articles/788898/',
-            title: 'Next.js App Router. Опыт использования. Путь в будущее или поворот не туда',
-            description: 'представила новый подход к роутингу, ' +
-                'который должен был стать заменой так называемому Pages Router, вместе с тем добавив ряд принципиально нового функционала.',
-            readTime: '3 минуты'
-        }
-    ]
 
+const BlogPage = async (Component) => {
     const posts: BlogPostsProps = await getPosts()
+    const imageUrl = 'https://habrastorage.org/r/w1560/getpro/habr/upload_files/da0/b0d/2cc/da0b0d2ccd24fe7ab50820c2803d3f9d.jpg'
 
-
-    console.log(posts)
     return (
         <div className={styles.cardWrapper}>
             {posts.map((item, id) => (
@@ -72,7 +21,7 @@ const BlogPage = async () => {
                     docsHrefText={item.docsHrefText}
                     tag={item.postId}
                     date={item.date}
-                    image={'https://habrastorage.org/r/w1560/getpro/habr/upload_files/da0/b0d/2cc/da0b0d2ccd24fe7ab50820c2803d3f9d.jpg'}
+                    image={imageUrl}
                     href={item.href}
                     title={item.name}
                     description={item.body}
