+++ app/components/Category/Category.tsx
@@ -4,13 +4,10 @@ import styles from './category.module.sass'
 import {P} from "@/app/components/P/P";
 import {ListTree} from "@/app/components/ListTree/ListTree";
 import React, {useState} from "react";
-import {ListTreeProps} from "@/app/components/ListTree/listTree.props";
 import {CategoryProps} from "@/app/components/Category/Category.props";
 import {ListProps} from "@/app/components/List/list.props";
 
-
 export const Category = ({data, name, children}: CategoryProps<ListProps>) => {
-
     const [isOpen, setIsOpen] = useState(false);
     const handleShowList = () => {
         setIsOpen(prevState => !prevState);
@@ -22,7 +19,5 @@ export const Category = ({data, name, children}: CategoryProps<ListProps>) => {
             <P className={styles.category_name} size={'l'}>{name}</P>
         </div>
         {isOpen && <ListTree data={data}/>}
-
-
     </div>
 }
