+++ app/components/List/List.tsx
@@ -4,6 +4,7 @@ import React from "react";
 import {ListTree} from "@/app/components/ListTree/ListTree";
 import {ListProps} from "@/app/components/List/list.props";
 import styles from "./list.module.sass"
+import Link from "next/link";
 
 export const List = ({_id, alias, category, title, pages}: ListProps) => {
     const [isOpen, setIsOpen] = React.useState(false);
@@ -16,7 +17,11 @@ export const List = ({_id, alias, category, title, pages}: ListProps) => {
     return <div onClick={(e) => {
         handleShowList(e)
     }}>
-        <a href={alias || '#'}>{_id.secondCategory || title}</a>
+
+        {/*<a href={alias || '#'}>{_id.secondCategory || title}</a>*/}
+        <Link href={alias ? `/courses/${alias}` : '#'}>
+            {_id.secondCategory || title}
+        </Link>
 
         {isOpen && <ListTree data={pages}/>}
     </div>
