+++ app/components/menu/Menu.module.css
@@ -10,10 +10,4 @@
   margin-left: 10px;
 }
 
-.hrefContainer {
-  margin-left: 20px;
-  display: flex;
-  flex-direction: column;
-}
-
 /*# sourceMappingURL=Menu.module.css.map */
