+++ app/components/menu/Menu.module.sass
@@ -8,8 +8,3 @@
   display: flex
   flex-direction: column
   margin-left: 10px
-
-.hrefContainer
-  margin-left: 20px
-  display: flex
-  flex-direction: column
