+++ app/courses/[alias]/page.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const Page = () => {
+    return (
+        <div>
+            adsasdad
+        </div>
+    );
+};
+
+export default Page;
